+++ internal_flash.h
@@ -22,7 +22,7 @@ uint16_t internalFlash_pageFromAddress(uint32_t address);
 uint32_t internalFlash_addressFromPage(uint16_t page);
 uint16_t internalFlash_addressWithinPage(uint32_t address, uint16_t page);
 
-uint8_t internalFlash_read(uint32_t address, uint8_t data_length, uint8_t* buffer);
+uint8_t internalFlash_read(uint32_t address, uint16_t data_length, uint8_t* buffer);
 uint8_t internalFlash_write(uint32_t address, uint8_t data_length, uint8_t* buffer);
